Merge main into sm

+++ BackendAPI/Controllers/RoomController.cs
@@ -73,6 +73,7 @@ namespace BackendAPI.Controllers
                 .Select(r => new RoomResult
                 {
                     Id = r.ID,
+                    HotelId = r.HotelID,
                     Name = r.Name,
                     Description = r.Description,
                     DailyPrice = r.DailyPrice,
@@ -97,23 +98,23 @@ namespace BackendAPI.Controllers
 
         [Authorize(Roles = "HotelAdmin")]
         [HttpPost("CreateRoom")]
-        public async Task<IActionResult> CreateRoom([FromForm] CreateRoomDTO roomDto)
+        public async Task<IActionResult> CreateRoom([FromBody] CreateRoomDTO roomDto)
         {
             var room = new Room()
             {
                 ID = Guid.NewGuid().ToString(),
                 HotelID = roomDto.HotelID,
                 DailyPrice = roomDto.DailyPrice,
-                CreatedAt = DateTime.UtcNow.AddHours(1),    
+                Name = roomDto.Name,
+                Description = roomDto.Description,
+                CreatedAt = DateTime.UtcNow.AddHours(1),
                 UpdatedAt = DateTime.UtcNow.AddHours(1),
             };
-
-           
 
             _context.Rooms.Add(room);
             await _context.SaveChangesAsync();
 
-            return Ok(room);
+            return Ok(new CreateRoomResult { Id = room.ID });
         }
 
         [Authorize(Roles = "HotelAdmin")]
@@ -304,11 +305,16 @@ namespace BackendAPI.Controllers
         {
             var room = await _context.Rooms.FindAsync(id);
 
+            if (room == null)
+            {
+                return NotFound($"Room with ID: {id} not found.");
+            }
+
             _context.Rooms.Remove(room);
 
             await _context.SaveChangesAsync();
 
-            return StatusCode(200,$"Room deleted succesfully {room}");
+            return StatusCode(200,$"Room deleted succesfully");
         }
